MIPS: fix generation of certain 64-bit constants
This fixes generation of 64-bit constants with equal high and low
halves and with bit 31 (and 63) set when using -G less than 8.

Constants of this form, when not loaded via gp, are constructed
using LUI/ORI for the low word, followed by a shift and or to fill
in the high word.  If bit 31 of this constant is set, the LUI
instruction sign extends setting all bits in the high word, and
these of course remain set after a bitwise or with the shifted low
word.

If the sign bit is set, we can instead construct the bitwise inverse
of the desired value and invert it back afterwards, thus avoiding
any sign extension issues.

To observe this behaviour, compile this function with -64 -G0:

unsigned long foo(void)
{
    return 0x8765432187654321;
}

Without this patch, the return value will be 0xffffffff87654321.